# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1",
    "title": "wallet",
    "description": ""
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/transactions": {
      "get": {
        "summary": "Get all transactions",
        "description": "Endpoint to get all transactions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Transactions retrieved"
          }
        }
      },
      "post": {
        "summary": "Create a new transaction",
        "description": "Endpoint to create a new transaction",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user": {
                    "type": "string",
                    "description": "User ID (references the 'User' model)"
                  },
                  "amount": {
                    "type": "number",
                    "description": "Transaction amount"
                  },
                  "category": {
                    "type": "string",
                    "description": "Transaction category",
                    "enum": ["category1", "category2"]
                  },
                  "date": {
                    "type": "string",
                    "description": "Transaction date"
                  },
                  "isIncome": {
                    "type": "boolean",
                    "description": "Indicates if the transaction is income"
                  },
                  "comment": {
                    "type": "string",
                    "description": "Transaction comment"
                  }
                },
                "required": ["user", "amount", "category", "date", "isIncome"]
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction created"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "SwaggerHub API Auto Mocking",
      "url": "https://virtserver.swaggerhub.com/wallet/"
    }
  ]
}